pyproject.toml
[project]
name = "htmx-counter-app"
version = "0.1.0"
description = "A simple counter application using FastAPI, HTMX, Pydantic, and Neon Serverless DB"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.27.0",
    "pydantic>=2.5.0",
    "jinja2>=3.1.2",
    "python-dotenv>=1.0.0",
    "psycopg2-binary>=2.9.9",
    "asyncpg>=0.29.0",
    "sqlalchemy>=2.0.0",
    "gymnasium>=0.29.1",
]

[project.optional-dependencies]
dev = [
    "black>=23.1.0",
    "isort>=5.12.0",
]

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = []
